chore(boardlist): created Cell-component and moved Row to BoardList/components/
