+++ next-tavla/src/Admin/scenarios/BoardList/components/Cell/styles.module.css
@@ -1,0 +1,4 @@
+.cell {
+    display: flex;
+    align-items: center;
+}
